+++ storage/core/jvm/src/main/scala/fluence/storage/InMemoryKVStore.scala
@@ -173,18 +173,4 @@ object InMemoryKVStore {
     protected val data: TrieMap[K, V] = map
   }
 
-  // this MonadError is needed for travers
-  implicit def storeMonadError[F[_]](implicit ME: MonadError[F, Throwable]): MonadError[F, StoreError] =
-    new MonadError[F, StoreError] {
-      override def flatMap[A, B](fa: F[A])(f: A ⇒ F[B]): F[B] = ME.flatMap(fa)(f)
-      override def tailRecM[A, B](a: A)(f: A ⇒ F[Either[A, B]]): F[B] = ME.tailRecM(a)(f)
-      override def raiseError[A](e: StoreError): F[A] = ME.raiseError(e)
-      override def handleErrorWith[A](fa: F[A])(f: StoreError ⇒ F[A]): F[A] = ME.handleErrorWith(fa) {
-        case cf: StoreError ⇒ f(cf)
-        case t ⇒ ME.raiseError(t)
-      }
-      override def pure[A](x: A): F[A] =
-        ME.pure(x)
-    }
-
 }
